requirements.txt
Fabric3
PrettyTable
paste
-r noodles/requirements.txt
future
